+++ src/app.js
@@ -5,11 +5,13 @@ import cors from "cors";
 import userRouter from "./router/user.router.js";
 import productsRouter from "./router/products.router.js";
 import adminRouter from "./router/admin.router.js";
+import marketRouter from "./router/market.router.js";
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
+app.use(marketRouter);
 app.use(productsRouter);
 app.use(userRouter);
 app.use(adminRouter);
